TIVI-1658: Fix command for adding the app user to the weston-launcher group
Using usermod will result in all the many aother critical groups that
the app user is already a part of being removed.  The app user is already
part of video and audio so the only thing needed is just to add it to
the one group which is want groupmod does.