ivi.yaml: enable quiet boot
Add the 'quiet' kernel option to make kernel boot quietly, which means that
kernel won't print hundreds of lines of messages to the console. This will make
it boot a lot faster.

The messages will be lated visible with 'dmesg' later on the console.

This _will_ make make our life more difficult in case of kernel issues, because
we won't see kernel error messages on the console. But we sacrifise this for
several seconds faster boot, which is extremely important in the IVI domain.

Change-Id: I9c822336f55df526a4157be9a5b419eda7ec20eb
Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>