Enable linger for the app user
Since user-session@5000.service does not work with systemd v208, and
Tizen is not using a display or session manager for user logins, a user
instance for "app" does not start up at all.

A workaround for this issue is to enable the "linger" for the app user
so that user@5000.service will launch at every boot.

The proper fix is to add a new service that will register a graphical
user session with systemd-logind, but this hasn't been implemented yet.
The feature is tracked in PTREL-521.

Change-Id: Ibce509523b4a6f2f158af07369823f8fde19f034
Signed-off-by: Patrick McCarty <patrick.mccarty@linux.intel.com>