ks: enable the 'fstab' installer framework plugin
Enable MIC's 'fstab' installer framework plugin. This will make sure MIC does
not generate '/etc/fstab' but leaves this task to the corresponding installer
framework plugin. The plugin is implemented as part of the 'setup-ivi'
projectm, see here for more details:

https://wiki.tizen.org/wiki/IVI/artem-setup-ivi

Note, this patch is part of a series of patches which switch Tizen IVI to
'setup-ivi'. The entire series has to be deployed at once.

Change-Id: I7dd979263063747055308d8e53f2f8c002204b86
Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>