ks: switch to setup-[extlinux|gummiboot]
Add the 'setup-gummiboot' package to EFI images and 'setup-extlinux' package to
MBR images. These packages are generated from the 'setup-ivi' project and
implement configuration file manipulation tools for gummiboot / extlinux.

This patch is part of a patch-set which switches Tizen IVI to the 'setup-ivi'
project. The whole patch-set has to be deployed at once.

Change-Id: Ib446931517d5d1da47d6ed38139f7aa9b72f1156
Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>