ks: finally switch to setup-ivi
This patch makes sure that all the KS files execute the 'setup-ivi-boot'
script, which is the installer framework plugin that installs and configures
the bootloader and generates the /etc/fstab file, see

https://wiki.tizen.org/wiki/IVI/artem-setup-ivi#Bootloader_setup

This patch removes the 'efi.post' and 'mbr.post' files, and adds a
'setup-ivi.post' file instead, because now we execute the same script for all
image types. The script has internal logic for distinguishing image types and
installing the right bootloader.

This patch requires 2 other preparational patches, and all 3 patches have to be
deployed at once.

Change-Id: I7a4cf826a70833c1b394a6399d05353cf51c81bb
Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>