partitions: use 'noatime' mount option explicitely
At the moment MIC is hard-coded to add the 'noatime' mount option to the
/etc/fstab file for all file-systems it generate. This also happens in case of
installer framework plugins: "noatime" is added to the
"INSTALLERFW_PARTx_FSOPTS" variable. However, this "featrue" has been fixed
recently and MIC will stop doing this soon for installer framework variables
(the legacy /etc/fstab generating code stays the same).

And since we are switching to 'setup-ivi-boot' installer framework plugin,
we will be affected by the change in MIC: newer MIC versions won't add
'noatime', but we probably want to have it.

Let's just add 'noatime' explicitely from the KS file using the
'--fsoptions=noatime' argument. With current versions of MIC we'll end up with
double 'noatime' in /etc/fstab, which is harmless. But once the new version of
MIC is deployed to the build servers, we'll have a single 'noatime' in
/etc/fstab.

Change-Id: Ibed24c3f05f54feb6cb02bd03583e844a5f57b37
Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>