Added filter definitions of image with build target
One new file was added: ivi-targets.yaml. Inside it, the relationship of
images and build targets is defined. It means which image belong to
which target, and the backend services will only generate the proper
images for specific target, to avoid current image duplication issue.

It need also cowork with the new update of `kickstarter` command, and
image-configurations build instruction. But this patch can be merged
before them, nothing will be affected.

Change-Id: I614ca6b410f09c7698e7658165fddfe845cf681d
Signed-off-by: JF Ding <jian-feng.ding@intel.com>